Split ctags command
